requirements.txt
streamlit>=1.31
requests>=2.31
python-dotenv>=1.0
pandas>=2.2
